chore: move test utils code to `tests/src`
