+++ packages/rolldown/tests/fixture.test.ts
@@ -1,5 +1,5 @@
 import { test } from 'vitest'
-import type { TestConfig } from './types'
+import type { TestConfig } from './src/types'
 import { InputOptions, OutputOptions, rolldown } from 'rolldown'
 import nodePath from 'node:path'
 import * as fastGlob from 'fast-glob'
