+++ packages/rolldown/tests/fixtures/input/object/_config.ts
@@ -1,7 +1,7 @@
 import type { RollupOptions, RollupOutput } from 'rolldown'
 import path from 'node:path'
 import { expect } from 'vitest'
-import { getOutputChunkNames } from '../../../util'
+import { getOutputChunkNames } from '@tests/util'
 
 const config: RollupOptions = {
   input: {
