+++ packages/rolldown/tsconfig.json
@@ -1,11 +1,9 @@
 {
   "include": [
     "src/**/*",
+    "tests/src/**/*",
     "tests/**/*.test.ts",
-    "tests/**/_config.ts",
-    "tests/index.ts",
-    "tests/types.ts",
-    "tests/utils.ts"
+    "tests/**/_config.ts"
   ],
   "exclude": ["src/binding.js"],
   "compilerOptions": {
@@ -39,8 +37,8 @@
     "moduleResolution": "Bundler" /* Specify how TypeScript looks up a file from a given module specifier. */,
     "baseUrl": "./" /* Specify the base directory to resolve non-relative module names. */,
     "paths": {
-      "@tests/*": ["tests/*"],
-      "@tests": ["tests/index.ts"]
+      "@tests/*": ["tests/src/*"],
+      "@tests": ["tests/src/index.ts"]
     } /* Specify a set of entries that re-map imports to additional lookup locations. */,
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
